server/ rotas de momories

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+    const { sub } = request.user
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,12 +28,90 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
-    const { id } = paramsSchema.parse(req.params)
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send({ error: 'Não autorizado' })
+    }
+    return memory
+  })
+
+  app.post('/memories', async (request) => {
+    const bodySchema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPublic: z.coerce.boolean().default(false),
+    })
+
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+
+    const memory = await prisma.memory.create({
+      data: {
+        content,
+        coverUrl,
+        isPublic,
+        userId: request.user.sub,
+      },
+    })
+
+    return memory
+  })
+
+  app.put('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    })
+
+    const { id } = paramsSchema.parse(request.params)
+
+    const bodySchema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPublic: z.coerce.boolean().default(false),
+    })
+
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({ error: 'Nao autorizado' })
+    }
+
+    memory = await prisma.memory.update({
+      where: {
+        id,
+      },
+      data: {
+        content,
+        coverUrl,
+        isPublic,
+      },
+    })
+
+    return memory
+  })
+
+  app.delete('/memories', async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    })
+
+    const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
@@ -32,65 +119,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
-    return memory
-  })
-
-  app.post('/memories', async (req) => {
-    const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    })
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
-
-    const memory = await prisma.memory.create({
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-        userId: '148f16af-3002-4a10-89b6-e7e4f1c9e8d3',
-      },
-    })
-
-    return memory
-  })
-
-  app.put('/memories/:id', async (req) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    })
-
-    const { id } = paramsSchema.parse(req.params)
-
-    const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    })
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
-
-    const memory = await prisma.memory.update({
-      where: {
-        id,
-      },
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-      },
-    })
-
-    return memory
-  })
-
-  app.delete('/memories', async (req) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    })
-
-    const { id } = paramsSchema.parse(req.params)
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({ error: 'Nao autorizado' })
+    }
 
     await prisma.memory.delete({
       where: {
